refactor: remove factory methods from main file and add to new factory file
